Refactored with files inside folders for better structuring of files and code.
